+++ apps/analytics/components/Table/GenericTable.tsx
@@ -13,12 +13,12 @@ interface GenericTableProps<C> {
   linkFormatter?(path: string): string
 }
 
-declare module '@tanstack/react-table' {
-  interface ColumnMeta<TData extends RowData, TValue> {
-    className?: string
-    skeleton: ReactNode
-  }
-}
+// declare module '@tanstack/react-table' {
+//   interface ColumnMeta<TData extends RowData, TValue> {
+//     className?: string
+//     skeleton: ReactNode
+//   }
+// }
 
 export const GenericTable = <T extends { id: string }>({
   table,
@@ -151,7 +151,7 @@ export const GenericTable = <T extends { id: string }>({
             {!loading && table.getRowModel().rows.length === 0 && (
               <Table.tr className="!h-[260px]">
                 <Table.td colSpan={table.getAllColumns().length} className="!h-[260px]">
-                  <Typography variant="xs" className="text-slate-400 italic w-full text-center">
+                  <Typography variant="xs" className="w-full italic text-center text-slate-400">
                     {placeholder}
                   </Typography>
                 </Table.td>