+++ apps/earn/components/PoolsSection/Tables/PositionsTable/PositionsTable.tsx
@@ -29,6 +29,8 @@ export const PositionsTable: FC = () => {
     (url) => fetch(url).then((response) => response.json())
   )
 
+  // console.log({userWithFarms})
+
   const table = useReactTable<UserWithFarm>({
     data: userWithFarms || [],
     state: {